refactor(shared-domain): Migrate enums and clean up deprecated value objects
Relocated `AccountStatus` and `MusicGenre` enums to core domain package. Removed obsolete value objects (`Location`, `InvalidLocationException`) and added factory methods (`of`, `reconstitute`) to existing value objects for better instantiation.
