+++ src/main/java/com/gresk/shared/domain/AccountStatus.java
@@ -1,4 +1,4 @@
-package com.gresk.shared.domain.valueobject;
+package com.gresk.shared.domain;
 
 public enum AccountStatus {
     PENDING,
