+++ src/main/java/com/gresk/shared/domain/Role.java
@@ -1,0 +1,5 @@
+package com.gresk.shared.domain;
+
+public enum Role {
+    USER, ARTIST, PROMOTER
+}
