--- src/main/java/com/gresk/shared/domain/exception/InvalidLocationException.java
@@ -1,7 +1,0 @@
-package com.gresk.shared.domain.exception;
-
-public class InvalidLocationException extends DomainException {
-    public InvalidLocationException(String message) {
-        super(message);
-    }
-}
