+++ src/main/java/com/gresk/shared/domain/valueobject/Email.java
@@ -20,6 +20,15 @@ public record Email(String value) {
         }
     }
 
+
+    public static Email of(String value) {
+        return new Email(value);
+    }
+
+    public static Email reconstitute(String value) {
+        return new Email(value);
+    }
+
     @Override
     public String toString() {
         return value;
